Add backend route for generating content.
